Use Language Name in `Code Example x` Headings
Convert all headings of the form `#### Code Example multLangIdentifier`
into `#### Code Example languageName`.
For this conversion each instance of `multLangIdentifier` is replaced by
the corresponding `languageName` as described in the `book.json` file
(in `.pluginsConfig.api-language-selector.languages`, where `lang` is
the `multLangIdentifier` and `name` is the `languageName`), which got
deleted with commit 499511e9ba03167575f2bd18accab0899cc60372.

Signed-off-by: Julian Schacher <jspp@posteo.net>